Methods to extract proximity matrices from random forests

Extracts proximity matrices from random forest objects from the party, randomForest or randomForestSRC packages

extract_proximity(fit, newdata)

Arguments

fit

object of class 'RandomForest', 'randomForest', or 'rfsrc'

newdata

new data with the same columns as the data used for fit

...

arguments to be passed to extract_proximity

Value

an n x n matrix where position i, j gives the proportion of times observation i and j are in the same teriminal node across all trees

See also

plot_prox for plotting principal components of proximity matrices.

Examples

fit <- randomForest(hp ~ ., mtcars, proximity = TRUE)
extract_proximity(fit)

#>                      Mazda RX4 Mazda RX4 Wag Datsun 710 Hornet 4 Drive
#> Mazda RX4           1.00000000    0.96078431 0.25531915     0.13114754
#> Mazda RX4 Wag       0.96078431    1.00000000 0.32142857     0.19402985
#> Datsun 710          0.25531915    0.32142857 1.00000000     0.22535211
#> Hornet 4 Drive      0.13114754    0.19402985 0.22535211     1.00000000
#>                     Hornet Sportabout    Valiant Duster 360  Merc 240D
#> Mazda RX4                  0.09836066 0.09836066 0.03921569 0.06779661
#> Mazda RX4 Wag              0.10714286 0.13207547 0.05000000 0.15492958
#> Datsun 710                 0.01724138 0.09375000 0.00000000 0.40000000
#> Hornet 4 Drive             0.21538462 0.68918919 0.04347826 0.40259740
#>                       Merc 230   Merc 280  Merc 280C Merc 450SE Merc 450SL
#> Mazda RX4           0.14285714 0.40540541 0.34328358 0.01754386 0.13114754
#> Mazda RX4 Wag       0.21666667 0.49019608 0.45901639 0.03278689 0.13333333
#> Datsun 710          0.46268657 0.14814815 0.17741935 0.02127660 0.01785714
#> Hornet 4 Drive      0.21875000 0.26086957 0.28915663 0.12903226 0.15384615
#>                     Merc 450SLC Cadillac Fleetwood Lincoln Continental
#> Mazda RX4            0.03333333         0.00000000          0.00000000
#> Mazda RX4 Wag        0.01515152         0.00000000          0.00000000
#> Datsun 710           0.01886792         0.00000000          0.00000000
#> Hornet 4 Drive       0.19696970         0.06349206          0.04761905
#>                     Chrysler Imperial   Fiat 128 Honda Civic Toyota Corolla
#> Mazda RX4                  0.01470588 0.01333333  0.01612903     0.01612903
#> Mazda RX4 Wag              0.01538462 0.03174603  0.04838710     0.01666667
#> Datsun 710                 0.00000000 0.16393443  0.32142857     0.13333333
#> Hornet 4 Drive             0.02739726 0.00000000  0.01449275     0.03030303
#>                     Toyota Corona Dodge Challenger AMC Javelin Camaro Z28
#> Mazda RX4              0.08823529       0.02857143  0.02941176 0.06153846
#> Mazda RX4 Wag          0.14705882       0.03174603  0.00000000 0.08064516
#> Datsun 710             0.54166667       0.00000000  0.01562500 0.00000000
#> Hornet 4 Drive         0.29032258       0.15189873  0.23333333 0.00000000
#>                     Pontiac Firebird  Fiat X1-9 Porsche 914-2 Lotus Europa
#> Mazda RX4                 0.05194805 0.04545455    0.33898305   0.05882353
#> Mazda RX4 Wag             0.02898551 0.01538462    0.35416667   0.05970149
#> Datsun 710                0.01515152 0.42857143    0.18181818   0.24590164
#> Hornet 4 Drive            0.22784810 0.01449275    0.00000000   0.07042254
#>                     Ford Pantera L Ferrari Dino Maserati Bora Volvo 142E
#> Mazda RX4               0.20634921   0.61111111    0.04687500 0.18181818
#> Mazda RX4 Wag           0.11940299   0.69491525    0.02857143 0.28358209
#> Datsun 710              0.00000000   0.20634921    0.00000000 0.56451613
#> Hornet 4 Drive          0.06250000   0.18644068    0.06578947 0.15277778
#>  [ reached getOption("max.print") -- omitted 28 rows ]
#> 

fit <- randomForest(Species ~ ., iris, proximity = TRUE)
extract_proximity(fit)

#>              1          2          3          4          5          6         7
#> 1   1.00000000 0.98437500 1.00000000 1.00000000 1.00000000 1.00000000 1.0000000
#>              8          9         10         11         12         13
#> 1   1.00000000 0.98750000 1.00000000 1.00000000 1.00000000 1.00000000
#>             14         15         16         17         18         19
#> 1   0.98461538 0.78481013 0.83561644 1.00000000 1.00000000 0.79120879
#>             20         21         22         23         24         25
#> 1   1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>             26         27         28         29         30         31
#> 1   1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#>             32        33         34         35         36         37         38
#> 1   1.00000000 1.0000000 0.82089552 1.00000000 1.00000000 0.87142857 1.00000000
#>             39         40         41         42         43        44         45
#> 1   1.00000000 1.00000000 1.00000000 0.95774648 1.00000000 1.0000000 1.00000000
#>             46         47         48         49         50         51
#> 1   0.98550725 1.00000000 1.00000000 1.00000000 1.00000000 0.00000000
#>             52         53         54         55         56         57
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>             58         59         60         61         62         63
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>             64         65         66         67         68         69
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>             70         71         72         73         74         75
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>             76         77         78         79         80         81
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>             82         83         84         85         86         87
#> 1   0.00000000 0.00000000 0.00000000 0.01369863 0.00000000 0.00000000
#>             88         89         90         91         92         93
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>             94         95         96         97         98         99
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            100        101        102        103        104        105
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            106        107        108        109        110        111
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            112        113        114        115        116        117
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            118        119        120        121        122        123
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            124        125        126        127        128        129
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            130        131        132        133        134        135
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            136        137        138        139        140        141
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            142        143        144        145        146        147
#> 1   0.00000000 0.00000000 0.00000000 0.00000000 0.00000000 0.00000000
#>            148        149        150
#> 1   0.00000000 0.00000000 0.00000000
#>  [ reached getOption("max.print") -- omitted 149 rows ]
#>