Variable importance using random forests

Computes local or aggregate variable importance for a set of predictors from a fitted random forest object from the party, randomForest, or randomForestSRC package

variable_importance(fit, var, type, interaction, oob, nperm, parallel, data)

Arguments

fit

object of class 'RandomForest', 'randomForest', or 'rfsrc'

var

character, variables to find the importance of

type

character, either "aggregate," (default) which gives the average loss under permutation or "local" which gives the residual across all observations

interaction

logcal, compute the joint and additive importance for observations (type = "local") or variables type = "aggregate"

oob

logical, use the out of bag data data

nperm

positive integer giving the number of times to permute the indicated variables (default 10)

parallel

logical whether to run in parallel using the registered backend (default FALSE)

data

optional (unless using randomForest) data.frame with which to calculate importance

Value

a named numeric vector for type = "aggregate" or a data.frame for type = "local"

See also

plot_imp for plotting the results of variable_importance.

Examples

library(randomForest)
data(iris)

fit <- randomForest(Species ~ ., iris)
variable_importance(fit, var = colnames(iris)[-5], type = "aggregate", nperm = 2, data = iris)

#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>  0.006666667  0.013333333  0.186666667  0.196666667 
#> attr(,"class")
#> [1] "importance" "numeric"   
#> attr(,"type")
#> [1] "aggregate"
#> attr(,"var")
#> [1] "Sepal.Length" "Sepal.Width"  "Petal.Length" "Petal.Width" 
#> attr(,"interaction")
#> [1] FALSE
#> attr(,"target")
#>          1          2          3          4          5          6          7 
#>     setosa     setosa     setosa     setosa     setosa     setosa     setosa 
#>          8          9         10         11         12         13         14 
#>     setosa     setosa     setosa     setosa     setosa     setosa     setosa 
#>         15         16         17         18         19         20         21 
#>     setosa     setosa     setosa     setosa     setosa     setosa     setosa 
#>         22         23         24         25         26         27         28 
#>     setosa     setosa     setosa     setosa     setosa     setosa     setosa 
#>         29         30         31         32         33         34         35 
#>     setosa     setosa     setosa     setosa     setosa     setosa     setosa 
#>         36         37         38         39         40         41         42 
#>     setosa     setosa     setosa     setosa     setosa     setosa     setosa 
#>         43         44         45         46         47         48         49 
#>     setosa     setosa     setosa     setosa     setosa     setosa     setosa 
#>         50         51         52         53         54         55         56 
#>     setosa versicolor versicolor versicolor versicolor versicolor versicolor 
#>         57         58         59         60         61         62         63 
#> versicolor versicolor versicolor versicolor versicolor versicolor versicolor 
#>         64         65         66         67         68         69         70 
#> versicolor versicolor versicolor versicolor versicolor versicolor versicolor 
#>         71         72         73         74         75         76         77 
#> versicolor versicolor versicolor versicolor versicolor versicolor versicolor 
#>         78         79         80         81         82         83         84 
#> versicolor versicolor versicolor versicolor versicolor versicolor versicolor 
#>         85         86         87         88         89         90         91 
#> versicolor versicolor versicolor versicolor versicolor versicolor versicolor 
#>         92         93         94         95         96         97         98 
#> versicolor versicolor versicolor versicolor versicolor versicolor versicolor 
#>         99        100        101        102        103        104        105 
#> versicolor versicolor  virginica  virginica  virginica  virginica  virginica 
#>        106        107        108        109        110        111        112 
#>  virginica  virginica  virginica  virginica  virginica  virginica  virginica 
#>        113        114        115        116        117        118        119 
#>  virginica  virginica  virginica  virginica  virginica  virginica  virginica 
#>        120        121        122        123        124        125        126 
#>  virginica  virginica  virginica  virginica  virginica  virginica  virginica 
#>        127        128        129        130        131        132        133 
#>  virginica  virginica  virginica  virginica  virginica  virginica  virginica 
#>        134        135        136        137        138        139        140 
#>  virginica  virginica  virginica  virginica  virginica  virginica  virginica 
#>        141        142        143        144        145        146        147 
#>  virginica  virginica  virginica  virginica  virginica  virginica  virginica 
#>        148        149        150 
#>  virginica  virginica  virginica 
#> Levels: setosa versicolor virginica
#> 

library(party)
data(swiss)
fit <- cforest(Fertility ~ ., swiss)
variable_importance(fit, var = colnames(swiss)[-1], type = "local", nperm = 2)

#>    Agriculture Examination    Education    Catholic Infant.Mortality
#> 1            0 -0.53163162  3.440268642  0.86369391     -0.092274045
#> 2            0 -3.18992211 -0.000427429 -0.79054502     -0.352885464
#> 3            0 -3.08921584 -1.405076674 -1.39515593     -0.352972676
#> 4            0 -0.92216271 -5.874165193 -2.35898749     -0.327574890
#> 5            0  1.64853249  5.986485555  0.76432590     -0.340166046
#> 6            0 -2.99337319 -2.294364244 -1.56128521     -0.215220971
#> 7            0 -1.65580077 -3.549619836 -1.37535940     -0.258775893
#> 8            0 -5.85151018 -3.078748716 -2.73775333     -0.036976288
#> 9            0  0.02463397 -2.316909922 -0.94186817     -0.431847533
#> 10           0 -4.65700001  5.550965828  0.00000000     -0.194647286
#> 11           0 -3.69389238 -3.756409256 -1.12259775     -0.427054068
#> 12           0  5.59681931  1.800417611 -0.03264072      0.563395753
#> 13           0 -1.96988862 -3.789269459  1.38473818      0.381685259
#> 14           0 -0.62024306  5.319656860  1.13455853     -0.301999896
#> 15           0  0.26584545 -6.518956521  1.87836743      0.198846836
#> 16           0  3.75330323 -3.067277653 -0.54728554      0.009802129
#> 17           0  1.49960453 -4.379777508  0.45626243     -0.005094077
#> 18           0  6.83836662  0.565706337  0.82200771     -0.007102162
#> 19           0  0.00000000  4.237756422  0.77026818      0.171341120
#> 20           0 -0.55856636  2.578576136  0.99688122      0.009939069
#> 21           0  3.10719311  2.456939615  0.86972166      0.471256283
#> 22           0 -2.59380669 -0.431705864  1.42848765     -0.242126419
#> 23           0  0.00000000  5.574088583  0.65708535      0.080071594
#> 24           0  2.96485804 -0.004300868  1.01709164      0.000000000
#> 25           0 -2.40403627 -2.321520139  1.38907366     -0.212456479
#> 26           0  0.13182844  0.422461895  2.53474167      0.000000000
#> 27           0 -3.29625391 -5.368266508  0.17056929      0.118407592
#> 28           0  1.56718165  1.785077380  0.12146993      0.000000000
#> 29           0  6.77851097  1.115868370  0.36517515     -0.087645377
#> 30           0 -4.94668378 -4.508529946  1.62639567     -0.112356306
#>  [ reached getOption("max.print") -- omitted 17 rows ]
#> 

data(mtcars)
library(randomForestSRC)
fit <- rfsrc(mpg ~ ., mtcars)
variable_importance(fit, var = colnames(mtcars)[-1], type = "aggregate", nperm = 2, oob = TRUE)

#>        cyl       disp         hp       drat         wt       qsec         vs 
#> 2.55142622 7.41557402 4.24586161 0.30047554 6.00988423 0.18548639 0.14559372 
#>         am       gear       carb 
#> 0.05118357 0.05343370 0.23097257 
#> attr(,"class")
#> [1] "importance" "numeric"   
#> attr(,"type")
#> [1] "aggregate"
#> attr(,"var")
#>  [1] "cyl"  "disp" "hp"   "drat" "wt"   "qsec" "vs"   "am"   "gear" "carb"
#> attr(,"interaction")
#> [1] FALSE
#> attr(,"target")
#>  [1] 21.0 21.0 22.8 21.4 18.7 18.1 14.3 24.4 22.8 19.2 17.8 16.4 17.3 15.2 10.4
#> [16] 10.4 14.7 32.4 30.4 33.9 21.5 15.5 15.2 13.3 19.2 27.3 26.0 30.4 15.8 19.7
#> [31] 15.0 21.4
#>